requirements.txt
discord.py
python-dotenv
aiofiles
pymongo
dnspython